+++ aries_vcx_core/src/wallet/mod.rs
@@ -2,3 +2,4 @@ pub mod agency_client_wallet;
 pub mod base_wallet;
 #[cfg(feature = "vdrtools")]
 pub mod indy_wallet;
+pub mod internal;
