+++ libvcx_core/src/api_vcx/api_global/wallet.rs
@@ -6,13 +6,12 @@ use aries_vcx::aries_vcx_core::anoncreds::base_anoncreds::BaseAnonCreds;
 #[cfg(all(feature = "anoncreds_credx"))]
 use aries_vcx::aries_vcx_core::anoncreds::credx_anoncreds::IndyCredxAnonCreds;
 use aries_vcx::aries_vcx_core::anoncreds::indy_anoncreds::IndySdkAnonCreds;
-use aries_vcx::aries_vcx_core::indy::wallet::{
-    close_search_wallet, fetch_next_records_wallet, import, open_search_wallet,
-};
+use aries_vcx::aries_vcx_core::indy::wallet::import;
 use aries_vcx::aries_vcx_core::wallet::base_wallet::BaseWallet;
 use aries_vcx::aries_vcx_core::wallet::indy_wallet::{IndySdkWallet, IssuerConfig, RestoreWalletConfigs, WalletConfig};
 use aries_vcx::aries_vcx_core::{indy, SearchHandle};
 use aries_vcx::aries_vcx_core::{INVALID_WALLET_HANDLE, WalletHandle};
+use aries_vcx::aries_vcx_core::wallet::internal::{close_search_wallet, fetch_next_records_wallet, open_search_wallet};
 use aries_vcx::common::signing::unpack_message_to_string;
 use aries_vcx::global::settings::DEFAULT_LINK_SECRET_ALIAS;
 use aries_vcx::protocols::mediated_connection::pairwise_info::PairwiseInfo;
